IDE: make alloy server mode configurable
