+++ legend-engine-core/legend-engine-core-pure/legend-engine-pure-ide/legend-engine-pure-ide-light-http-server/src/main/java/org/finos/legend/engine/ide/ServerConfiguration.java
@@ -28,6 +28,7 @@ public class ServerConfiguration extends Configuration
     public SwaggerBundleConfiguration swagger;
     public SourceLocationConfiguration sourceLocationConfiguration;
     public Boolean debugMode;
+    public Boolean alloyServerMode;
     public Boolean initPureRuntimeBeforeStart;
     public List<String> requiredRepositories;
     public List<String> excludeRepositories;
